services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
pydantic==2.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
aiosmtplib==3.0.1
email-validator==2.1.0
authlib==1.3.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# MinIO / Object Storage
minio==7.2.0

# Image Processing
Pillow==10.1.0

# Cryptography (for token encryption)
cryptography==41.0.7